changes to route for login Repair_Sys, css file altered
for login route, included usage of bcrypt to handle login authentication,  instead of checking for matching username and passwords we use queries to check if username exists and if the password entered matches the database that stores hashed passwords. additionally we have added in a logger for flask applications. In the CSS file we have altered the appearance of borders, backround colours, alignment of divs and font sizes for home page. Additionally i have altered the font size for all h1 and h2 tags in all routes