updating template webpage ,css file and Repair_Sys
login page: altered number of characters for password.
Layout webpage: created alert-div as a container for alerts in navigation bar
home webpage: altered script eventlistener and added a div for modal, this feature must still be altered. altered script for conditional statement to alter the pseudo after element and display the alert correctly when Repair orders are due.
in CSS file, added styling for alert-div, which is the container for alerts in navigation bar. Added styling for this alert-div for the after psuedo element.
Repair_Sys file , altered home page variable msg  to customerdue.